organized components for reusability
